# openapi.yaml
openapi: 3.0.3
info:
  title: Claims Management System API
  description: |-
    API specification for the Claims Management System, a Spring Boot application for managing insurance claims.
    
    ## Features
    - Create and manage claimants
    - Submit and track insurance claims
    - Assign claims to staff members
    - Monitor workload and claims status
    
    ## Claim Lifecycle States
    - SUBMITTED: Initial claim submission
    - ASSIGNED: Claim assigned to a staff member
    - ASSESSMENT_IN_PROGRESS: Under review
    - MORE_INFO_REQUESTED: Additional information needed
    - APPROVED: Claim approved
    - REJECTED: Claim rejected
    - SETTLEMENT_IN_PROGRESS: Settlement being processed
    - CLOSED: Claim closed
  contact:
    name: API Support
    email: support@chubb.com
  version: 0.0.1-SNAPSHOT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.claimsmanagement.com
    description: Production server
...
paths:
...
  /api/v1/staff/{staffNumber}:
    get:
      tags:
        - Staff
      summary: Get a staff member
      description: Retrieve a staff member by ID
      operationId: getStaffMember
      parameters:
        - name: staffNumber
          in: path
          required: true
          schema:
            type: string
          description: Public staff number
      responses:
        '200':
          description: Staff member found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStaff'
        '404':
          description: Staff member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
...
  /api/v1/staff/{staffNumber}/claims/queue:
    get:
      tags:
        - Staff
      summary: Get claims picked up by staff
      description: Retrieve claims currently picked up by a staff member
      operationId: getStaffClaimQueue
      parameters:
        - name: staffNumber
          in: path
          required: true
          schema:
            type: string
          description: Public staff number
      responses:
        '200':
          description: Staff queue retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStaffClaimQueueList'
        '404':
          description: Staff member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /api/v1/staff/{staffNumber}/claims/queue/{claimNumber}/pickup:
    post:
      tags:
        - Staff
      summary: Pick up a claim
      description: Atomically assign an available claim to a staff member
      operationId: pickUpClaim
      parameters:
        - name: staffNumber
          in: path
          required: true
          schema:
            type: string
        - name: claimNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Claim picked up successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStaffClaimQueue'
        '400':
          description: Claim cannot be picked up or is owned by another staff member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: Staff member or available claim not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /api/v1/staff/{staffNumber}/claims/queue/{claimNumber}/requeue:
    post:
      tags:
        - Staff
      summary: Requeue a picked-up claim
      description: Return a claim to the available queue
      operationId: requeueClaim
      parameters:
        - name: staffNumber
          in: path
          required: true
          schema:
            type: string
        - name: claimNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Claim requeued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStaffClaimQueue'
        '400':
          description: Claim is assigned to another staff member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: Staff member or picked-up claim not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
...
components:
  schemas:
...
    ClaimType:
      type: string
      enum:
        - MOTOR
        - PROPERTY
      description: |-
        Type of insurance claim
        - MOTOR: Motor vehicle insurance claim
        - PROPERTY: Property damage insurance claim
      example: MOTOR
    ClaimStatus:
      type: string
      enum:
        - SUBMITTED
        - ASSIGNED
        - ASSESSMENT_IN_PROGRESS
        - MORE_INFO_REQUESTED
        - APPROVED
        - REJECTED
        - SETTLEMENT_IN_PROGRESS
        - CLOSED
      description: |-
        Current status of the claim in its lifecycle
        - SUBMITTED: Initial claim submission
        - ASSIGNED: Claim assigned to staff member
        - ASSESSMENT_IN_PROGRESS: Under assessment
        - MORE_INFO_REQUESTED: Awaiting additional information
        - APPROVED: Claim approved
        - REJECTED: Claim rejected
        - SETTLEMENT_IN_PROGRESS: Settlement being processed
        - CLOSED: Claim closed
      example: SUBMITTED
...
    Staff:
      type: object
      properties:
        staffNumber:
          type: string
          description: Public staff number
          example: STAFF-1001
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      description: Staff member who can process claims
    StaffClaimQueue:
      type: object
      properties:
        claimNumber:
          type: string
          description: Public claim number
          example: CLM-ABC12345
        claimantMemberNumber:
          type: string
        claimType:
          $ref: '#/components/schemas/ClaimType'
        claimStatus:
          $ref: '#/components/schemas/ClaimStatus'
        queueStatus:
          type: string
          enum:
            - AVAILABLE
            - PICKED_UP
        staffNumber:
          type: string
          description: Public staff number of the current owner
          example: STAFF-1001
          nullable: true
        queuedAt:
          type: string
          format: date-time
        pickedUpAt:
          type: string
          format: date-time
          nullable: true
      description: A claim waiting for or currently assigned to a staff member
    ApiResponseStaff:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Staff'
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ApiResponseStaffClaimQueue:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/StaffClaimQueue'
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ApiResponseStaffClaimQueueList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/StaffClaimQueue'
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ApiResponseError:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: false
        data:
          type: 'null'
          description: No data in error response
        message:
          type: string
          description: Error message
          example: "Resource not found"
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the response
          example: "2024-01-15T10:30:00.000Z"
      description: API response wrapper for error responses
...